Simplifying platforms
Each of its function is now in a single source file
Makefile updated this way
\#ifdef includes for paltform specific control
