+++ src/linux.h
@@ -11,6 +11,18 @@
 #include <ncurses.h>
 #include <unistd.h>
 
+#define PLAYGROUND_X 35
+#define PLAYGROUND_Y 20
+#define DEFAULT_SPEED 100
+#define TIME clock()
+#define DEFAULT_MOVE_TIME 5
+#include <ncurses.h>
+
+#define DEFAULT_LENGTH 5
+
+
+void init_platform();
+void gracefully_shutdown(char* message);
 
 void print_at(int x, int y, char * text);
 // to get a character but non blocking
