--- src/numworks/init.c
@@ -1,33 +1,0 @@
-/*
-
-init.c
-
-function for init and deinit
-
-*/
-
-#include <stdio.h>
-#include <stdlib.h>
-#include <time.h>
-
-
-
-
-void quit(int type)
-{
-	#if defined(__linux__)
-		#include <ncurses.h>
-		endwin();
-	#endif
-    switch (type) {
-        case 1:
-            printf("\nOuch!\n");
-            break; 
-        case 2: 
-            printf("\nGood Job!\n");
-            break;
-        case 3:
-            printf("\nBye bye !\n\n");
-    }
-    exit(0);
-}
