+++ src/windows.h
@@ -4,12 +4,22 @@
  *
  * */
 
-#ifndef PLATFORM_H
-#define PLATFORM_H
+#ifndef WINDOWS_H
+#define WINDOWS_H
 
 #include "main.h"
-#include <ncurses.h>
 #include <unistd.h>
+
+#define PLAYGROUND_X 35
+#define PLAYGROUND_Y 20
+#define DEFAULT_SPEED 100
+#define TIME clock()
+#define DEFAULT_MOVE_TIME 5
+#define DEFAULT_LENGTH 5
+
+
+void init_platform();
+void gracefully_shutdown(char* message);
 
 
 void print_at(int x, int y, char * text);
